feat(TodoStore.ts): add TodoStore to handle CRUD operations for todos
The TodoStore file is added to the TodoFrontend/src/stores directory. This store is responsible for handling CRUD operations for todos. It defines the following interfaces: Todo and TodoCreation. The Todo interface represents the structure of a todo item, including its id, title, content, and done status. The TodoCreation interface represents the structure of a new todo item, including its title and content.

The useTodoStore function is exported from the file using the defineStore function from the pinia library. This function defines the actions that can be performed on the todo store. The actions include:

- getMany: Fetches multiple todos from the server using the TodoServices.GetMany endpoint.
- getOne: Fetches a single todo by its id from the server using the TodoServices.GetOne endpoint.
- create: Creates a new todo on the server using the TodoServices.Create endpoint.
- updateState: Updates the done status of a todo on the server using the TodoServices.Update endpoint.
- delete: Deletes a todo from the server using the TodoServices.Delete endpoint.

Each action makes an HTTP request to the server using the fetch function. If the request is not successful (response.ok is false), an error is thrown. Otherwise, the response data is returned or the action is completed.

The TodoServices module is imported from '@/services/Todo.Service' to provide the endpoints for the todo API.